pyproject.toml
[project]
name = "mrkt-screener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "lxml>=5.3.1",
    "lxml-html-clean>=0.4.1",
    "marketprofile>=0.2.0",
    "matplotlib>=3.10.1",
    "mplfinance>=0.12.10b0",
    "nbformat>=5.10.4",
    "newspaper3k>=0.2.8",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
    "tabulate>=0.9.0",
    "textblob>=0.19.0",
    "yfinance>=0.2.54",
    "flake8>=7.1.2",
    "tqdm>=4.67.1",
    "tenacity>=9.0.0",
]
